feat: reorganize tests and move spaLoginRequiredView to separate module
