+++ backend/backend/urls.py
@@ -17,7 +17,6 @@ Including another URLconf
 from django.contrib import admin
 from django.urls import path, include
 from drf_spectacular.views import SpectacularAPIView, SpectacularRedocView, SpectacularSwaggerView
-from .tests import spaLoginRequiredView
 
 urlpatterns = [
     path('api/admin/', admin.site.urls),
@@ -25,5 +24,5 @@ urlpatterns = [
     path('api/schema/', SpectacularAPIView.as_view(), name='schema'),
     path('api/schema/swagger-ui/', SpectacularSwaggerView.as_view(url_name='schema'), name='swagger-ui'),
     path('api/schema/redoc/', SpectacularRedocView.as_view(url_name='schema'), name='redoc'),
-    path('login_required/', spaLoginRequiredView, name='login_required'),
+    path('tests/', include('tests.urls')),
 ]
