+++ backend/tests/apps.py
@@ -1,0 +1,6 @@
+from django.apps import AppConfig
+
+
+class TestsConfig(AppConfig):
+    default_auto_field = 'django.db.models.BigAutoField'
+    name = 'tests'
